+++ src/main/java/io/github/astrapi69/crypt/api/compound/package-info.java
@@ -1,4 +1,0 @@
-/**
- * This package provides an enum class with compound java security standard algorithms
- */
-package io.github.astrapi69.crypt.api.compound;