+++ src/test/java/io/github/astrapi69/crypt/api/algorithm/MdAlgorithmTest.java
@@ -26,8 +26,13 @@ package io.github.astrapi69.crypt.api.algorithm;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
+import io.github.astrapi69.crypt.api.algorithm.compound.CompoundAlgorithm;
+import io.github.astrapi69.crypt.api.algorithm.compound.CompoundAlgorithmTest;
 import org.junit.jupiter.api.Test;
 
+/**
+ * The test class {@link MdAlgorithmTest} provides unit tests for the {@link MdAlgorithm} enum.
+ */
 public class MdAlgorithmTest
 {
 	/**