+++ src/test/java/io/github/astrapi69/crypt/api/algorithm/compound/CompoundAlgorithmTest.java
@@ -28,6 +28,9 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import org.junit.jupiter.api.Test;
 
+/**
+ * The test class {@link CompoundAlgorithmTest} provides unit tests for the {@link CompoundAlgorithm} enum.
+ */
 public class CompoundAlgorithmTest
 {