+++ CHANGELOG.md
@@ -4,7 +4,9 @@
 Version 9.2-SNAPSHOT
 -------------
 
+CHANGED:
 
+- remove of deprecated enum class CompoundAlgorithm
 
 Version 9.1
 -------------